pyproject.toml
[project]
name = "silver-disco"
version = "0.1.0"
description = "Add your description here"
authors = [{ name = "afaulconbridge" }]
dependencies = [
    "imutils>=0.5.4",
    "ffmpeg-python>=0.2.0",
    "opencv-contrib-python-headless>=4.9.0.80",
    "matplotlib>=3.9.0",
]
readme = "README.md"
requires-python = ">= 3.12"

[project.scripts]
"silver-disco" = "silver_disco:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = ["pytest>=8.1.1", "pre-commit>=3.7.0", "ruff>=0.3.5"]

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/silver_disco"]

[tool.ruff]
# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".ipynb_checkpoints",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pyenv",
    ".pytest_cache",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    ".vscode",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "site-packages",
    "venv",
]

# Same as Black.
line-length = 88
indent-width = 4

# Assume Python 3.12
target-version = "py312"

[tool.ruff.lint]
ignore = [
    "COM812",
    "COM819",
    "D100",
    "D101",
    "D102",
    "D103",
    "D104",
    "D105",
    "D107",
    "D203",
    "D213",
    "D300",
    "E111",
    "E114",
    "E117",
    "ISC001",
    "ISC002",
    "S311",
    "Q000",
    "Q001",
    "Q002",
    "Q003",
    "W191",
    "ANN101",
    "ANN204",
]
select = ["ALL"]

# Enable Pyflakes (`F`) and a subset of the pycodestyle (`E`)  codes by default.
# Allow fix for all enabled rules (when `--fix`) is provided.
fixable = ["ALL"]
unfixable = []

# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.format]
# Like Black, use double quotes for strings.
quote-style = "double"

# Like Black, indent with spaces, rather than tabs.
indent-style = "space"

# Like Black, respect magic trailing commas.
skip-magic-trailing-comma = false

# Like Black, automatically detect the appropriate line ending.
line-ending = "auto"
